+++ cupy/_lib/_routines_poly.py
