+++ cupy/_lib/stride_tricks.py
