--- cupy/lib/__init__.py
@@ -1,8 +1,0 @@
-from cupy.lib import stride_tricks  # NOQA
-from cupy.lib.polynomial import poly1d  # NOQA
-from cupy.lib._routines_poly import polyadd  # NOQA
-from cupy.lib._routines_poly import polysub  # NOQA
-from cupy.lib._routines_poly import polymul  # NOQA
-from cupy.lib._routines_poly import polyfit  # NOQA
-from cupy.lib._routines_poly import polyval  # NOQA
-from cupy.lib._routines_poly import roots  # NOQA
